Fix TypeScript build errors for Vercel deploy
- Add missing imports for saveUserToStorage and isTelegramUser functions
- Fix Telegram WebApp type conflicts by consolidating types in src/types/telegram.ts
- Add missing properties (active, createdAt) to Task objects in store
- Replace incorrect Lucide icon 'Tasks' with 'CheckSquare'
- Fix react-hot-toast type error by replacing toast.info with toast()
- Remove unused imports causing TypeScript warnings
- Resolve initDataUnsafe type conflicts by using type assertions

All TypeScript errors resolved - build now passes successfully.
