+++ src/components/TasksPanel.tsx
@@ -2,7 +2,6 @@ import React, { useState } from 'react';
 import { motion } from 'framer-motion';
 import { 
   CheckSquare, 
-  Clock, 
   Users, 
   Calendar, 
   Trophy,
