+++ src/types/telegram.ts
@@ -1,0 +1,15 @@
+export interface TelegramUser {
+  id: number;
+  first_name: string;
+  last_name?: string;
+  username?: string;
+  photo_url?: string;
+  language_code?: string;
+}
+
+export interface TelegramWebAppInitData {
+  user?: TelegramUser;
+  start_param?: string;
+  auth_date: number;
+  hash: string;
+}
